# scrape: Speed up scrape once we have reached sample_limit