@roidelapluie 2020-01-28
Note that it will no longer convert the samples to series and therefore the count will also contain the series that are in in droppedSeries.

As we are in a case of breaking sample limits anyway, I think it is acceptable as we want to quickly finish the scrape.