@roidelapluie 2020-01-28
I tested manually with a 5.000.000 /metrics endpoint with a limit at 5000

![5000000-4999-2](https://user-images.githubusercontent.com/291750/73307846-55418c80-421f-11ea-8115-4752f622146f.png)

![5000000-4999](https://user-images.githubusercontent.com/291750/73307870-62f71200-421f-11ea-8947-af1878352938.png)

scrape time goes to 5.5s to 0.8s