@roidelapluie 2020-01-28
Benchmark

```
benchmark                                              old ns/op       new ns/op       delta
BenchmarkScrapeLoopAppendSampleLimit/100x1.1-8         43556           13052           -70.03%
BenchmarkScrapeLoopAppendSampleLimit/100x1.5-8         59586           17147           -71.22%
BenchmarkScrapeLoopAppendSampleLimit/100x2-8           77491           23140           -70.14%
BenchmarkScrapeLoopAppendSampleLimit/100x5-8           200492          53887           -73.12%
BenchmarkScrapeLoopAppendSampleLimit/100x10-8          416163          107292          -74.22%
BenchmarkScrapeLoopAppendSampleLimit/100x100-8         3866957         1026895         -73.44%
BenchmarkScrapeLoopAppendSampleLimit/100x1000-8        37042364        10557452        -71.50%
BenchmarkScrapeLoopAppendSampleLimit/100x10000-8       355821480       105005649       -70.49%
BenchmarkScrapeLoopAppendSampleLimit/1000x1.1-8        455672          118611          -73.97%
BenchmarkScrapeLoopAppendSampleLimit/1000x1.5-8        629227          162510          -74.17%
BenchmarkScrapeLoopAppendSampleLimit/1000x2-8          777553          219727          -71.74%
BenchmarkScrapeLoopAppendSampleLimit/1000x5-8          1869539         534025          -71.44%
BenchmarkScrapeLoopAppendSampleLimit/1000x10-8         4112363         1062212         -74.17%
BenchmarkScrapeLoopAppendSampleLimit/1000x100-8        38682654        10414268        -73.08%
BenchmarkScrapeLoopAppendSampleLimit/1000x1000-8       373562383       102323049       -72.61%
BenchmarkScrapeLoopAppendSampleLimit/1000x10000-8      3488088420      1053417567      -69.80%
BenchmarkScrapeLoopAppendSampleLimit/10000x1.1-8       4451121         1169922         -73.72%
BenchmarkScrapeLoopAppendSampleLimit/10000x1.5-8       5887881         1658950         -71.82%
BenchmarkScrapeLoopAppendSampleLimit/10000x2-8         8155631         2182022         -73.25%
BenchmarkScrapeLoopAppendSampleLimit/10000x5-8         20445595        5400332         -73.59%
BenchmarkScrapeLoopAppendSampleLimit/10000x10-8        39684982        10706167        -73.02%
BenchmarkScrapeLoopAppendSampleLimit/10000x100-8       357654689       104841981       -70.69%
BenchmarkScrapeLoopAppendSampleLimit/10000x1000-8      3656158681      1094190744      -70.07%
BenchmarkScrapeLoopAppendSampleLimit/10000x10000-8     34971398423     10529485385     -69.89%

benchmark                                              old allocs     new allocs     delta
BenchmarkScrapeLoopAppendSampleLimit/100x1.1-8         452            8              -98.23%
BenchmarkScrapeLoopAppendSampleLimit/100x1.5-8         612            8              -98.69%
BenchmarkScrapeLoopAppendSampleLimit/100x2-8           812            8              -99.01%
BenchmarkScrapeLoopAppendSampleLimit/100x5-8           2012           8              -99.60%
BenchmarkScrapeLoopAppendSampleLimit/100x10-8          4012           8              -99.80%
BenchmarkScrapeLoopAppendSampleLimit/100x100-8         40012          8              -99.98%
BenchmarkScrapeLoopAppendSampleLimit/100x1000-8        400012         12             -100.00%
BenchmarkScrapeLoopAppendSampleLimit/100x10000-8       4000015        42             -100.00%
BenchmarkScrapeLoopAppendSampleLimit/1000x1.1-8        4412           8              -99.82%
BenchmarkScrapeLoopAppendSampleLimit/1000x1.5-8        6012           8              -99.87%
BenchmarkScrapeLoopAppendSampleLimit/1000x2-8          8012           8              -99.90%
BenchmarkScrapeLoopAppendSampleLimit/1000x5-8          20012          9              -99.96%
BenchmarkScrapeLoopAppendSampleLimit/1000x10-8         40012          11             -99.97%
BenchmarkScrapeLoopAppendSampleLimit/1000x100-8        400012         48             -99.99%
BenchmarkScrapeLoopAppendSampleLimit/1000x1000-8       4000016        409            -99.99%
BenchmarkScrapeLoopAppendSampleLimit/1000x10000-8      40000024       4020           -99.99%
BenchmarkScrapeLoopAppendSampleLimit/10000x1.1-8       44012          48             -99.89%
BenchmarkScrapeLoopAppendSampleLimit/10000x1.5-8       60012          60             -99.90%
BenchmarkScrapeLoopAppendSampleLimit/10000x2-8         80012          81             -99.90%
BenchmarkScrapeLoopAppendSampleLimit/10000x5-8         200012         185            -99.91%
BenchmarkScrapeLoopAppendSampleLimit/10000x10-8        400013         368            -99.91%
BenchmarkScrapeLoopAppendSampleLimit/10000x100-8       4000019        4010           -99.90%
BenchmarkScrapeLoopAppendSampleLimit/10000x1000-8      40000034       40029          -99.90%
BenchmarkScrapeLoopAppendSampleLimit/10000x10000-8     400000033      40029          -99.99%

benchmark                                              old bytes       new bytes     delta
BenchmarkScrapeLoopAppendSampleLimit/100x1.1-8         11864           320           -97.30%
BenchmarkScrapeLoopAppendSampleLimit/100x1.5-8         16024           320           -98.00%
BenchmarkScrapeLoopAppendSampleLimit/100x2-8           21224           320           -98.49%
BenchmarkScrapeLoopAppendSampleLimit/100x5-8           52426           321           -99.39%
BenchmarkScrapeLoopAppendSampleLimit/100x10-8          104428          322           -99.69%
BenchmarkScrapeLoopAppendSampleLimit/100x100-8         1040473         338           -99.97%
BenchmarkScrapeLoopAppendSampleLimit/100x1000-8        10400851        532           -99.99%
BenchmarkScrapeLoopAppendSampleLimit/100x10000-8       104004181       2093          -100.00%
BenchmarkScrapeLoopAppendSampleLimit/1000x1.1-8        114858          340           -99.70%
BenchmarkScrapeLoopAppendSampleLimit/1000x1.5-8        156470          347           -99.78%
BenchmarkScrapeLoopAppendSampleLimit/1000x2-8          208488          360           -99.83%
BenchmarkScrapeLoopAppendSampleLimit/1000x5-8          520596          409           -99.92%
BenchmarkScrapeLoopAppendSampleLimit/1000x10-8         1040759         493           -99.95%
BenchmarkScrapeLoopAppendSampleLimit/1000x100-8        10403471        2286          -99.98%
BenchmarkScrapeLoopAppendSampleLimit/1000x1000-8       104031461       19987         -99.98%
BenchmarkScrapeLoopAppendSampleLimit/1000x10000-8      1040093232      196992        -99.98%
BenchmarkScrapeLoopAppendSampleLimit/10000x1.1-8       1151819         3377          -99.71%
BenchmarkScrapeLoopAppendSampleLimit/10000x1.5-8       1570982         4295          -99.73%
BenchmarkScrapeLoopAppendSampleLimit/10000x2-8         2094094         5906          -99.72%
BenchmarkScrapeLoopAppendSampleLimit/10000x5-8         5235436         13800         -99.74%
BenchmarkScrapeLoopAppendSampleLimit/10000x10-8        10471692        27646         -99.74%
BenchmarkScrapeLoopAppendSampleLimit/10000x100-8       104664960       303643        -99.71%
BenchmarkScrapeLoopAppendSampleLimit/10000x1000-8      1041993888      3033552       -99.71%
BenchmarkScrapeLoopAppendSampleLimit/10000x10000-8     10401993808     3033536       -99.97%
```